// web/src/api/generated/models/ActivityItem.ts
export interface ActivityItem {
    /**
     * 
     * @type {AuthorSummary}
     * @memberof ActivityItem
     */
    actor?: AuthorSummary | null;
    /**
     * 
     * @type {string}
     * @memberof ActivityItem
     */
    actorId: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ActivityItem
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ActivityItem
     */
    id: string;
    /**
     * 
     * @type {ActivityKind}
     * @memberof ActivityItem
     */
    kind: ActivityKind;
    /**
     * 
     * @type {Date}
     * @memberof ActivityItem
     */
    occurredAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ActivityItem
     */
    sourceEventId: string;
    /**
     * 
     * @type {ActivityTargetPresentation}
     * @memberof ActivityItem
     */
    target?: ActivityTargetPresentation | null;
    /**
     * 
     * @type {string}
     * @memberof ActivityItem
     */
    targetId: string | null;
    /**
     * 
     * @type {EngagementTarget}
     * @memberof ActivityItem
     */
    targetType: EngagementTarget | null;
}