// web/src/api/generated/models/DashboardItem.ts
export interface DashboardItem {
    /**
     * 
     * @type {Date}
     * @memberof DashboardItem
     */
    createdAt: Date | null;
    /**
     * 
     * @type {string}
     * @memberof DashboardItem
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof DashboardItem
     */
    occurredOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof DashboardItem
     */
    previewAttachmentId?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof DashboardItem
     */
    scheduledAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof DashboardItem
     */
    scheduledOn?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof DashboardItem
     */
    titleOrText: string | null;
    /**
     * 
     * @type {DashboardItemType}
     * @memberof DashboardItem
     */
    type: DashboardItemType;
}